# backend/Api/openapi.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "My Title",
    "version": "1.0.0"
  },
  "paths": {
...
    "/exhibit/accept/{id}": {
      "post": {
        "tags": [
          "Exhibit"
        ],
        "operationId": "Exhibit_AcceptConnection",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/exhibit/forget/{id}": {
      "post": {
        "tags": [
          "Exhibit"
        ],
        "operationId": "Exhibit_ForgetConnection",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
...
    "/exhibit/details/{exhibit_id}": {
      "get": {
        "tags": [
          "Exhibit"
        ],
        "operationId": "Exhibit_GetExhibit",
        "parameters": [
          {
            "name": "exhibit_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExhibitProperties"
                }
              }
            }
          }
        }
      }
    },
    "/exhibit/send/{exposition_id}": {
      "post": {
        "tags": [
          "Exhibit"
        ],
        "operationId": "Exhibit_SendPackages",
        "parameters": [
          {
            "name": "exposition_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ExhibitProperties": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "hostname": {
            "type": "string",
            "nullable": true
          },
          "tags": {
            "type": "array",
            "nullable": true,
            "items": {
              "type": "string"
            }
          },
          "sensors": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/SensorProperties"
            }
          }
        }
      },
      "SensorProperties": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "path": {
            "type": "string",
            "nullable": true
          },
          "friendlyName": {
            "type": "string",
            "nullable": true
          },
          "valueType": {
            "$ref": "#/components/schemas/ValueType"
          },
          "availableEvents": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "ValueType": {
        "type": "integer",
        "description": "",
        "x-enumNames": [
          "Void",
          "Bool",
          "Number",
          "Event",
          "Complex"
        ],
        "enum": [
          0,
          1,
          2,
          3,
          4
        ]
      },
...
    }
  }
}